+++ app/src/components/visualizers/sphere/base.tsx
@@ -11,12 +11,12 @@ import {
   COORDINATE_TYPE,
   type ICoordinateMapper,
   TWO_PI,
-} from "../../mappers/coordinateMappers/common";
+} from "@/lib/mappers/coordinateMappers/common";
 import {
   ColorPalette,
   type ColorPaletteType,
   COLOR_PALETTE,
-} from "../palettes";
+} from "@/lib/palettes";
 
 // const MAPPING_MODE_POLAR_2D = "polar_2d";
 // const MAPPING_MODE_POLAR_PHI = "polar_phi";
