+++ app/src/lib/analyzers/common.ts
