+++ app/src/lib/analyzers/fft.ts
