+++ app/src/lib/analyzers/scope.ts
