+++ app/src/lib/appState.ts
