+++ app/src/lib/applicationModes.ts
