+++ app/src/lib/mappers/valueTracker/common.ts
