+++ app/src/main.tsx
@@ -2,12 +2,12 @@ import { Leva } from "leva";
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 
-import App from "./App";
-import "./style/globals.css";
-import { CameraControlsContextProvider } from "./context/cameraControls";
-import { ModeContextProvider } from "./context/mode";
-import { ThemeProvider } from "./context/theme";
-import { VisualContextProvider } from "./context/visual";
+import App from "@/App";
+import "@/style/globals.css";
+import { CameraControlsContextProvider } from "@/context/cameraControls";
+import { ModeContextProvider } from "@/context/mode";
+import { ThemeProvider } from "@/context/theme";
+import { VisualContextProvider } from "@/context/visual";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
